requirements.txt
anthropic>=0.34.2
openai>=1.42.0
pyyaml>=6.0.2
python-dotenv>=1.0.1
requests>=2.32.3
beautifulsoup4>=4.12.3
pydantic>=2.9.2
httpx>=0.27.2
markdown>=3.7
google-api-python-client>=2.151.0
google-auth>=2.36.0
google-auth-httplib2>=0.2.0
pillow>=10.4.0
azure-storage-blob>=12.19.0
